modified existining scripts

+++ 00_Basic/02_Basics/03_Practical/kph_to_mph_converter.py
@@ -2,6 +2,6 @@
 """
 conversion formula for kph to mph is : 1 kilometre = 0.621371192 milesconversion formula for kph to mph is : 1 kilometre = 0.621371192 miles
 """
-kmh = int(raw_input("Enter km/h: "))
+kmh = float(raw_input("Enter km/h: "))
 mph =  0.6214 * kmh
 print "Speed:", kmh, "KM/H = ", mph, "MPH"
